Move waitlist data store to store module
