+++ network-protection/network-protection-impl/src/main/java/com/duckduckgo/networkprotection/impl/waitlist/store/NetPWaitlistRepository.kt
@@ -20,6 +20,7 @@ import com.duckduckgo.app.di.AppCoroutineScope
 import com.duckduckgo.app.global.DispatcherProvider
 import com.duckduckgo.di.scopes.AppScope
 import com.duckduckgo.networkprotection.impl.state.NetPFeatureRemover
+import com.duckduckgo.networkprotection.store.waitlist.NetPWaitlistDataStore
 import com.squareup.anvil.annotations.ContributesBinding
 import com.squareup.anvil.annotations.ContributesMultibinding
 import javax.inject.Inject
